+++ src/Header.js
@@ -1,82 +1,71 @@
-import AppBar from '@material-ui/core/AppBar';
-import Grid from '@material-ui/core/Grid';
-import { makeStyles } from '@material-ui/core/styles';
-import Toolbar from '@material-ui/core/Toolbar';
 import React from 'react';
-import { Typography } from '@material-ui/core';
-// import kpr100 from './001_UPDATE-01.png';
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
+import { makeStyles } from '@material-ui/core/styles';
 
-const useStyles = makeStyles({
+const useStyles = makeStyles((theme) => ({
     root: {
-        display: 'flex',
-        minWidth: 1280,
-        maxWidth: 1000,
-        flexDirection: 'column',
-        // flexGrow: 1,
-        padding: '20px',
-        // width: '1250px',
+        flexGrow: 1,
+        paddingRight: theme.spacing(11),
+        paddingLeft: theme.spacing(11),
     },
     logo: {
-        display: 'flex',
-        // justifyContent: 'flex-end',
+        // marginRight: theme.spacing(15),
+        width: 150,
+        height: 150,
+        marginLeft: 30,
+        marginRight: 20,
+        marginTop: 20,
+        marginBottom: 20,
+    },
+    toolbar: {
+        flexGrow: 1,
+        // borderRadius: 50,
+        backgroundColor: '#757DE8',
+        // borderRadius: 5,
+        // minHeight: 120,
+        // minWidth: 120,
+        // alignItems: 'flex-end',
+        alignItems: 'center',
+        paddingTop: theme.spacing(2),
+        paddingBottom: theme.spacing(2),
+    },
+    title: {
+        flexGrow: 1,
+        // display: 'none',
+        marginLeft: 220,
+        marginRight: 20,
+        textAlign: 'center',
+        [theme.breakpoints.up('sm')]: {
+            display: 'block',
+        },
         // alignSelf: 'flex-end',
-        width: 135,
-        height: 135,
     },
-    header: {
-        display: 'block',
-        // display: 'flex',
-        // flexWrap: 'wrap',
-        fontSize: '30px',
-    },
-});
+}));
 
-const Header = () => {
+export default function SearchAppBar() {
     const classes = useStyles();
 
     return (
-        <nav className={classes.root}>
-            <AppBar
-                position="static"
-                color="inherit"
-                style={{
-                    backgroundColor: '#757DE8',
-                    borderRadius: '3px',
-                    fontSize: '15px',
-                }}
-            >
-                <Toolbar
-                    style={{
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                    }}
-                >
-                    <Grid justify={'center'} alignItems={'center'} container>
-                        <Grid item>
-                            <Grid container justify={'center'}>
-                                <Typography
-                                    className={classes.header}
-                                    // variant="body3"
-                                    // color="textPrimary"
-                                    // component="h1"
-                                >
-                                    EMOTIONAL RESCUE COMMUNICATION & CONNECTION
-                                    GAMES
-                                </Typography>
-                            </Grid>
-                        </Grid>
-                        <Grid style={{ justifySelf: 'center' }} item>
-                            <img
-                                className={classes.logo}
-                                src={'./001_UPDATE-01.png'}
-                                alt="Keep It Real 100 Logo"
-                            />
-                        </Grid>
-                    </Grid>
+        <div className={classes.root}>
+            <AppBar position="static">
+                <Toolbar className={classes.toolbar}>
+                    <Typography
+                        className={classes.title}
+                        variant="h4"
+                        color="textPrimary"
+                        // noWrap
+                    >
+                        EMOTIONAL RESCUE COMMUNICATION & CONNECTION GAMES
+                    </Typography>
+                    <img
+                        className={classes.logo}
+                        src={'./001_UPDATE-01.png'}
+                        alt="Keep It Real 100 Logo"
+                    />
                 </Toolbar>
             </AppBar>
-        </nav>
+        </div>
     );
-};
-
-export default Header;
+}
